creating aimec out and cleaning
